// src/app/shared/models/sample-combination.ts
export class SampleCombinationDto {
    companyId: number;
    sampleId: number;
    sampleItem: number;
    combinationTypeId: number;
    parentSize: string;
    size: string;
    colour: string;
    quantity: number;
    unitMeasureId: string;
    salesUnit: string;
    unitPrice: number;
    salePrice: number;
    requestDate: string;
    internalProductId: string;
    externalProductId: string;
    sampleStatusId: number | null;
    createdByUser: string;
    creationDate: string;
    modifiedByUser: string;
    modifiedDate: string;
  }
  